requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.6
django-bootstrap-v5==1.0.7
django-chartjs==2.2.1
importlib-metadata==3.10.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
zipp==3.6.0
